+++ pygaia/errors/astrometric.py
@@ -1,7 +1,7 @@
 __all__ = ['parallaxErrorSkyAvg', 'parallaxErrorSkyAvgAltStartGate', 'positionErrorSkyAvg',
            'properMotionErrorSkyAvg', 'parallaxError', 'positionError', 'properMotionError']
 
-from numpy import sqrt, sin, array, floor
+from numpy import sqrt, sin, array, floor, power
 from scipy import isscalar
 from pygaia.errors.utils import calcZ, calcZAltStartGate
 from numpy import genfromtxt
@@ -21,6 +21,9 @@ _scalingForProperMotions = {'Total':0.526, 'AlphaStar':0.556, 'Delta':0.496}
 
 # The upper band of the parallax errors at the bright end
 _parallaxErrorMaxBright = 14.0
+
+# The nominal mission lifetime
+_nominalLifeTime = 5.0
 
 def errorScalingFactor(observable, beta):
   """
@@ -50,7 +53,26 @@ def errorScalingFactor(observable, beta):
     indices[(indices==_numStepsSinBeta)] = _numStepsSinBeta-1
     return _astrometricErrorFactors[observable][indices]
 
-def parallaxErrorSkyAvg(G, vmini):
+def errorScalingMissionLength(extension, p):
+    """
+    Calculate the factor by which to scale errors for a given Gaia mission extension.
+
+    Parameters
+    ----------
+
+    extension - The mission extension in years (a negative extension can be used to make crude
+    performance predictions part-way through the mission lifetime).
+    p - The power by which the errors scale with time (error ~ t^p, p=-0.5 for parallax and celestial
+    position, and -1.5 for proper motion).
+
+    Returns
+    -------
+
+    The factor by which to scale the errors
+    """
+    return power((_nominalLifeTime+extension)/_nominalLifeTime, p)
+
+def parallaxErrorSkyAvg(G, vmini, extension=0.0):
   """
   Calculate the sky averaged parallax error from G and (V-I).
 
@@ -60,15 +82,21 @@ def parallaxErrorSkyAvg(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The parallax error in micro-arcseconds.
   """
+  factor = errorScalingMissionLength(extension, -0.5)
   z=calcZ(G)
-  return sqrt(-1.631 + 680.766*z + 32.732*z*z)*(0.986 + (1.0 - 0.986)*vmini)
+  return sqrt(-1.631 + 680.766*z + 32.732*z*z)*(0.986 + (1.0 - 0.986)*vmini)*factor
 
-def parallaxMinError(G, vmini):
+def parallaxMinError(G, vmini, extension=0.0):
   """
   Calculate the minimum parallax error from G and (V-I). This correspond to the sky regions with the
   smallest astrometric errors.  At the bright end the parallax error is at least 14 muas due to the
@@ -80,14 +108,19 @@ def parallaxMinError(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The minimum parallax error in micro-arcseconds.
   """
-  return _astrometricErrorFactors["parallax"].min()*parallaxErrorSkyAvg(G, vmini)
+  return _astrometricErrorFactors["parallax"].min()*parallaxErrorSkyAvg(G, vmini, extension=extension)
 
-def parallaxMaxError(G, vmini):
+def parallaxMaxError(G, vmini, extension=0.0):
   """
   Calculate the maximum parallax error from G and (V-I). This correspond to the sky regions with the
   largest astrometric errors.  At the bright end the parallax error is at least 14 muas due to the
@@ -99,17 +132,22 @@ def parallaxMaxError(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The maximum parallax error in micro-arcseconds.
   """
-  errors = _astrometricErrorFactors["parallax"].max()*parallaxErrorSkyAvg(G, vmini)
+  errors = _astrometricErrorFactors["parallax"].max()*parallaxErrorSkyAvg(G, vmini, extension=extension)
   indices = (errors<_parallaxErrorMaxBright)
   errors[indices]=_parallaxErrorMaxBright
   return errors
 
-def parallaxError(G, vmini, beta):
+def parallaxError(G, vmini, beta, extension=0.0):
   """
   Calculate the parallax error from G and (V-I) and the Ecliptic latitude beta of the source. The
   parallax error is calculated by applying numerical factors to the sky average error. These factors
@@ -125,14 +163,19 @@ def parallaxError(G, vmini, beta):
   vmini - Value(s) of (V-I) colour.
   beta  - Value(s) of the Ecliptic latitude.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The parallax error in micro-arcseconds.
   """
-  return parallaxErrorSkyAvg(G, vmini)*errorScalingFactor('parallax',beta)
+  return parallaxErrorSkyAvg(G, vmini, extension=extension)*errorScalingFactor('parallax',beta)
 
-def parallaxErrorSkyAvgAltStartGate(G, vmini):
+def parallaxErrorSkyAvgAltStartGate(G, vmini, extension=0.0):
   """
   Calculate the sky averaged parallax error from G and (V-I). In this case assume gating starts at G=13.3
   (to simulate bright star worst performance)
@@ -143,15 +186,21 @@ def parallaxErrorSkyAvgAltStartGate(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The parallax error in micro-arcseconds.
   """
+  factor = errorScalingMissionLength(extension, -0.5)
   z=calcZAltStartGate(G)
-  return sqrt(-1.631 + 680.766*z + 32.732*z*z)*(0.986 + (1.0 - 0.986)*vmini)
+  return sqrt(-1.631 + 680.766*z + 32.732*z*z)*(0.986 + (1.0 - 0.986)*vmini)*factor
 
-def positionErrorSkyAvg(G, vmini):
+def positionErrorSkyAvg(G, vmini, extension=0.0):
   """
   Calculate the sky averaged position errors from G and (V-I).
 
@@ -164,16 +213,21 @@ def positionErrorSkyAvg(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The error in alpha* and the error in delta, in that order, in micro-arcsecond.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  parallaxError = parallaxErrorSkyAvg(G, vmini, extension=extension)
   return _scalingForPositions['AlphaStar']*parallaxError, \
          _scalingForPositions['Delta']*parallaxError
 
-def positionMinError(G, vmini):
+def positionMinError(G, vmini, extension=0.0):
   """
   Calculate the minimum position errors from G and (V-I). These correspond to the sky regions with the
   smallest astrometric errors.
@@ -187,16 +241,21 @@ def positionMinError(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The minimum error in alpha* and the error in delta, in that order, in micro-arcsecond.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  parallaxError = parallaxErrorSkyAvg(G, vmini, extension=extension)
   return _astrometricErrorFactors['alphaStar'].min()*parallaxError, \
          _astrometricErrorFactors['delta'].min()*parallaxError
 
-def positionMaxError(G, vmini):
+def positionMaxError(G, vmini, extension=0.0):
   """
   Calculate the maximum position errors from G and (V-I). These correspond to the sky regions with the
   largest astrometric errors.
@@ -210,16 +269,21 @@ def positionMaxError(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The maximum error in alpha* and the error in delta, in that order, in micro-arcsecond.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  parallaxError = parallaxErrorSkyAvg(G, vmini, extension)
   return _astrometricErrorFactors['alphaStar'].max()*parallaxError, \
          _astrometricErrorFactors['delta'].max()*parallaxError
 
-def positionError(G, vmini, beta):
+def positionError(G, vmini, beta, extension=0.0):
   """
   Calculate the position errors from G and (V-I) and the Ecliptic latitude beta of the source.
 
@@ -233,16 +297,21 @@ def positionError(G, vmini, beta):
   vmini - Value(s) of (V-I) colour.
   beta  - Value(s) of the Ecliptic latitude.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The error in alpha* and the error in delta, in that order, in micro-arcsecond.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  parallaxError = parallaxErrorSkyAvg(G, vmini, extension=extension)
   return errorScalingFactor('alphaStar',beta)*parallaxError, \
          errorScalingFactor('delta',beta)*parallaxError
 
-def properMotionErrorSkyAvg(G, vmini):
+def properMotionErrorSkyAvg(G, vmini, extension=0.0):
   """
   Calculate the sky averaged proper motion errors from G and (V-I).
 
@@ -255,16 +324,22 @@ def properMotionErrorSkyAvg(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The error in mu_alpha* and the error in mu_delta, in that order, in micro-arcsecond/year.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  factor = errorScalingMissionLength(extension, -1.5)
+  parallaxError = parallaxErrorSkyAvg(G, vmini)*factor
   return _scalingForProperMotions['AlphaStar']*parallaxError, \
          _scalingForProperMotions['Delta']*parallaxError
 
-def properMotionMinError(G, vmini):
+def properMotionMinError(G, vmini, extension=0.0):
   """
   Calculate the minimum proper motion errors from G and (V-I). These correspond to the sky regions with
   the smallest astrometric errors.
@@ -278,16 +353,22 @@ def properMotionMinError(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The minimum error in mu_alpha* and the error in mu_delta, in that order, in micro-arcsecond/year.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  factor = errorScalingMissionLength(extension, -1.5)
+  parallaxError = parallaxErrorSkyAvg(G, vmini)*factor
   return _astrometricErrorFactors['muAlphaStar'].min()*parallaxError, \
          _astrometricErrorFactors['muDelta'].min()*parallaxError
 
-def properMotionMaxError(G, vmini):
+def properMotionMaxError(G, vmini, extension=0.0):
   """
   Calculate the maximum proper motion errors from G and (V-I). These correspond to the sky regions with
   the largest astrometric errors.
@@ -301,18 +382,24 @@ def properMotionMaxError(G, vmini):
   G     - Value(s) of G-band magnitude.
   vmini - Value(s) of (V-I) colour.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The maximum error in mu_alpha* and the error in mu_delta, in that order, in micro-arcsecond/year.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  factor = errorScalingMissionLength(extension, -1.5)
+  parallaxError = parallaxErrorSkyAvg(G, vmini)*factor
   indices = (parallaxError<_parallaxErrorMaxBright)
   parallaxError[indices] = _parallaxErrorMaxBright
   return _astrometricErrorFactors['muAlphaStar'].max()*parallaxError, \
          _astrometricErrorFactors['muDelta'].max()*parallaxError
 
-def properMotionError(G, vmini, beta):
+def properMotionError(G, vmini, beta, extension=0.0):
   """
   Calculate the proper motion errors from G and (V-I) and the Ecliptic latitude beta of the source.
 
@@ -326,11 +413,46 @@ def properMotionError(G, vmini, beta):
   vmini - Value(s) of (V-I) colour.
   beta  - Value(s) of the Ecliptic latitude.
 
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
   Returns
   -------
 
   The error in mu_alpha* and the error in mu_delta, in that order, in micro-arcsecond/year.
   """
-  parallaxError = parallaxErrorSkyAvg(G, vmini)
+  factor = errorScalingMissionLength(extension, -1.5)
+  parallaxError = parallaxErrorSkyAvg(G, vmini)*factor
   return errorScalingFactor('muAlphaStar',beta)*parallaxError, \
          errorScalingFactor('muDelta',beta)*parallaxError
+
+def totalProperMotionErrorSkyAvg(G, vmini, extension=0.0):
+  """
+  Calculate the sky averaged total proper motion error from G and (V-I). This refers to the error on the
+  length of the proper motion vector.
+
+  NOTE! THE ERRORS ARE FOR PROPER MOTIONS IN THE ICRS (I.E., RIGHT ASCENSION, DECLINATION). MAKE SURE
+  YOUR SIMULATED ASTROMETRY IS ALSO ON THE ICRS.
+
+  Parameters
+  ----------
+
+  G     - Value(s) of G-band magnitude.
+  vmini - Value(s) of (V-I) colour.
+
+  Keywords
+  --------
+
+  extension - Add this amount of years to the mission lifetime and scale the errors accordingly.
+
+  Returns
+  -------
+
+  The error on the total proper motion in micro-arcsecond/year.
+  """
+  factor = errorScalingMissionLength(extension, -1.5)
+  parallaxError = parallaxErrorSkyAvg(G, vmini)*factor
+  return _scalingForProperMotions['Total']*parallaxError
+